+++ src/scrapers/movie/HotMovies.cpp
@@ -18,6 +18,7 @@ HotMovies::HotMovies(QObject* parent) :
         MovieScraperInfo::Runtime,
         MovieScraperInfo::Overview,
         MovieScraperInfo::Poster,
+        MovieScraperInfo::Backdrop,
         MovieScraperInfo::Actors,
         MovieScraperInfo::Genres,
         MovieScraperInfo::Studios,
@@ -153,7 +154,7 @@ void HotMovies::parseAndAssignInfos(QString html, Movie* movie, QSet<MovieScrape
     QRegExp rx;
     rx.setMinimal(true);
 
-    rx.setPattern(R"(<h1 class="title" itemprop="name">(.*)</h1>)");
+    rx.setPattern(R"(<h1 class="title"(?: itemprop="name")?>(.*)</h1>)");
     if (infos.contains(MovieScraperInfo::Title) && rx.indexIn(html) != -1) {
         movie->setName(rx.cap(1));
     }
@@ -175,12 +176,12 @@ void HotMovies::parseAndAssignInfos(QString html, Movie* movie, QSet<MovieScrape
         movie->ratings().push_back(rating);
     }
 
-    rx.setPattern("<span itemprop=\"copyrightYear\">([0-9]{4})</span>");
+    rx.setPattern("<strong>Released:</strong> ?([0-9]{4})");
     if (infos.contains(MovieScraperInfo::Released) && rx.indexIn(html) != -1) {
         movie->setReleased(QDate::fromString(rx.cap(1), "yyyy"));
     }
 
-    rx.setPattern(R"(<span itemprop="duration" datetime="PT[^"]+">(.*)</span>)");
+    rx.setPattern(R"(<span(?: itemprop="duration")? datetime="PT[^"]+">([^<]*)</span>)");
     if (infos.contains(MovieScraperInfo::Runtime) && rx.indexIn(html) != -1) {
         using namespace std::chrono;
         QStringList runtimeStr = rx.cap(1).split(":");
@@ -194,7 +195,7 @@ void HotMovies::parseAndAssignInfos(QString html, Movie* movie, QSet<MovieScrape
         }
     }
 
-    rx.setPattern(R"(<span class="video_description" itemprop="description">(.*)</span>)");
+    rx.setPattern(R"(<span class="video_description"(?: itemprop="description")?>(.*)</span>)");
     if (infos.contains(MovieScraperInfo::Overview) && rx.indexIn(html) != -1) {
         QTextDocument doc;
         doc.setHtml(rx.cap(1));
@@ -205,12 +206,20 @@ void HotMovies::parseAndAssignInfos(QString html, Movie* movie, QSet<MovieScrape
         }
     }
 
-    rx.setPattern(R"rx(<img itemprop="image" alt="[^"]*" id="cover"[\s\n]*[^>]*src="([^"]*)")rx");
+    rx.setPattern(R"rx(data-front="([^"]+)")rx");
     if (infos.contains(MovieScraperInfo::Poster) && rx.indexIn(html) != -1) {
         Poster p;
         p.thumbUrl = rx.cap(1);
         p.originalUrl = rx.cap(1);
         movie->images().addPoster(p);
+    }
+
+    rx.setPattern(R"rx(data-back="([^"]+)")rx");
+    if (infos.contains(MovieScraperInfo::Backdrop) && rx.indexIn(html) != -1) {
+        Poster p;
+        p.thumbUrl = rx.cap(1);
+        p.originalUrl = rx.cap(1);
+        movie->images().addBackdrop(p);
     }
 
     if (infos.contains(MovieScraperInfo::Actors)) {
@@ -225,7 +234,8 @@ void HotMovies::parseAndAssignInfos(QString html, Movie* movie, QSet<MovieScrape
 
         //        rx.setPattern("<div class=\"star_wrapper\" key=\"(.*)\"><a href=\".*\" .* title=\".*\" rel=\"tag\"
         //        itemprop=\"url\"><span itemprop=\"name\">(.*)</span></a></div>");
-        rx.setPattern(R"re(<div class="star_wrapper" key="([^"]*)"><img .*/><span itemprop="name">([^<]*)</span>)re");
+        rx.setPattern(
+            R"re(<div class="star_wrapper" key="([^"]*)"><img [^>]*/><span(?: itemprop="name")?>([^<]*)</span>)re");
         rx.setMinimal(true);
         int offset = 0;
         while ((offset = rx.indexIn(html, offset)) != -1) {
@@ -241,26 +251,20 @@ void HotMovies::parseAndAssignInfos(QString html, Movie* movie, QSet<MovieScrape
     }
 
     if (infos.contains(MovieScraperInfo::Genres)) {
-        rx.setPattern("<span itemprop=\"genre\">.* -> (.*)</span>");
+        rx.setPattern("title=\"Plot Oriented -> ([^\"]+)\"");
         int offset = 0;
         while ((offset = rx.indexIn(html, offset)) != -1) {
             offset += rx.matchedLength();
-            // Some "genres" are just some categories of HotMovies
-            const auto genre = rx.cap(1);
-            if (genre != "Streaming Video" && genre != "Downloads") {
-                movie->addGenre(genre);
-            }
+            movie->addGenre(rx.cap(1));
         }
     }
 
-    rx.setPattern("<strong>Studio:</strong> <a itemprop=\"url\" href=\"[^\"]*\"[\\s\\n]*title=\"[^\"]*\"><span "
-                  "itemprop=\"name\">(.*)</span></a>");
+    rx.setPattern(R"re(<strong>Studio:</strong> <a(?: itemprop="url")? href="[^"]*"[\s\t\n]*title="([^"]*)")re");
     if (infos.contains(MovieScraperInfo::Studios) && rx.indexIn(html) != -1) {
         movie->addStudio(rx.cap(1));
     }
 
-    rx.setPattern(R"(<span itemprop="director" itemscope itemtype="http://schema.org/Person"><a itemprop="url" )"
-                  R"(href="[^"]*"[\s\n]*title="[^"]*" rel="tag"><span itemprop="name">([^<]*)</span></a>)");
+    rx.setPattern(R"re("director":\[\{"@type":"Person","name":"([^"]+)")re");
     if (infos.contains(MovieScraperInfo::Director) && rx.indexIn(html) != -1) {
         movie->setDirector(rx.cap(1));
     }